ROX-13175: add table to show baseline flows
